Revisited the Combat System
This is a pretty humbling committ; I had to completely rework the combat
system due to the issues involving the Outposts. The combat system, as
previously designed, wasn't robust enough to handle battles that need to
be paused and restarted. This has now be addressed and only needs minor
tweaking to be sufficient.