Fixed Combat Bugs
Work was mainly spent cleaning up the combat flow and event system.
Solved the MissingReferenceExceptions by unsubscribing to events when
the object is slated for destruction. The curFights list is now being
culled outside of the BattleCompleted event method in CombatManager.
This was causing issues with the foreach used in it's Update method.
Finally, outpostSpawner is correctly detecting units inside its spawn
area and not spawning more. Even though it was fun to watch a unit get
catapulted down the lane when the spawn area wasn't clear.