Building the TeamGameObject
Began work on the TeamGameObject that manages the activities of each
team for the game. First to be completed where the Treasury and
PurchaseList components. The Treasury handles a team's funding and the
other is a collection of units that can be purchased. Next to be added
is the component to handle the spawning of the units.