Added TeamSpawner
Added the TeamSpawner behaviour for the TeamGameObject and got it
connected with the outposts and the Treasury behaviour. With a simple
change in the input manager, units now spawn in for the player (the Blue
Team) and deduct from their available funds, based on the PurchaseList.
Next goal is to add in an basic HUD for important information and a UI
interface for the outposts to allow for additional units to be spawned.